% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMin.R
\name{checkMin}
\alias{checkMin}
\title{Generates the lowest points along the paths}
\usage{
checkMin(
  dsm,
  tracks,
  export = TRUE,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05,
  st_dev = 0.06
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{Digital Surface Model raster file as '.tif'.}

\item{export}{If 'TRUE' the GeoPackage will be exported to your wd.}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}

\item{st_dev}{Minimum standarddeviation of crossprofile dsm-value 'z'. Acts as a filter. Defaults to '0.06'.}
}
\value{
Saves a GeoPackage in the outputfolder of the minimumpoints
}
\description{
Can save a GeoPackage and return a sf object of minimumpoints.
}
