% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPolygon.R
\name{checkPolygon}
\alias{checkPolygon}
\title{Generates a polygon around the input lines as wide as the detected width of the track}
\usage{
checkPolygon(
  dsm,
  tracks,
  export = FALSE,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05,
  st_dev = 0.06
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{Digitized tracks as GeoPackage '.gpkg'.}

\item{export}{If 'TRUE' the GeoPackage will be exported to your wd.}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}

\item{st_dev}{Minimum standarddeviation of crossprofile dsm-value 'z'. Acts as a filter. Defaults to '0.06'.}
}
\value{
a buffer around the input lines representing the calculated width of the tracks. It includes a column with the width in meters.
}
\description{
Generates a polygon around the input lines as wide as the detected width of the track
}
\examples{
polygon_as_width <- checkPolygon(dsm , tracks, export = FALSE, dist_cross = 1,profile_length = 1, dist_cross_points = 0.05,st_dev = 0.06
polygon_as_width
}
