% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRtrack-package.R
\docType{package}
\name{checkRtrack-package}
\alias{checkRtrack}
\alias{checkRtrack-package}
\title{checkRtrack: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Leonie Sonntag \email{leonie.sonntag@stud-mail.uni-wuerzburg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
