% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSides.R
\name{checkSides}
\alias{checkSides}
\title{Generates the track limit on the right and left side of the track}
\usage{
checkSides(
  dsm,
  tracks,
  export = TRUE,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{Digital Surface Model raster file as '.tif'.}

\item{export}{If 'TRUE' the GeoPackage will be exported to your wd.}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}
}
\description{
Can save a GeoPackage and return a sf object of right and left side points
}
\examples{
dsm <- read_dsm(system.file("tif/dsm.tif", package = "checkRtrack"))
tracks <- read_tracks(system.file("geopackage/tracks.gpkg", package = "checkRtrack"))
sides <- checkSides(dsm, tracks, export = FALSE, dist_cross = 1, profile_length = 1, dist_cross_points = 0.05)
str(sides)

checkMap(dsm, tracks, points= sides) #run checkMap() as many times as you like to get different examples

}
