% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWidth.R
\name{checkWidth}
\alias{checkWidth}
\title{Generates the average width of the tracks.}
\usage{
checkWidth(
  dsm,
  tracks,
  export = FALSE,
  plot = TRUE,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05,
  st_dev = 0.06
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{Digitized tracks as GeoPackage '.gpkg'.}

\item{export}{If 'TRUE' the GeoPackage will be exported to your wd.}

\item{plot}{If 'TRUE' a plot will be generated.}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}

\item{st_dev}{Minimum standarddeviation of crossprofile dsm-value 'z'. Acts as a filter. Defaults to '0.06'.}
}
\value{
The width information in meters added to the imported tracks objects attribute table and a plot
of the width vs. the track type (labelled).
}
\description{
It is recommended to use the other check-functions first. You can plot and export your track widths. The width is defined here as the distance of the steepest points from each side of the track calculated with cross profiles.
}
\examples{
dsm <- read_dsm(system.file("tif/dsm.tif", package = "checkRtrack"))
tracks <- read_tracks(system.file("geopackage/tracks.gpkg", package = "checkRtrack"))
width <- checkWidth(dsm , tracks, export = FALSE, plot = TRUE, dist_cross = 1, profile_length = 1, dist_cross_points = 0.05, st_dev = 0.06)
width

}
