% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptrack.R
\name{maptrack}
\alias{maptrack}
\title{Creating an overview map of a random scene of the data to check the results
A follow up after using function 'min_points' and 'outer_points'. Their
output files will be used here.
An overview op of the entire tracks and points would be too messy
so each time running this function, a random spot on your tracks will be mapped to see how it worked out.
The idea is to adjust the parameters used in the functions that created the points.}
\usage{
maptrack(dsm, minimumpoints, tracks, side1_extent = NULL, side2_extent = NULL)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{minimumpoints}{Output file of function 'min_points' as GeoPackage.}

\item{tracks}{Digital Surface Model raster file as '.tif'.}

\item{side1_extent}{Output file of the function 'outer_points'. Left.}

\item{side2_extent}{Output file of the function 'outer_points'. Right.}
}
\value{
A plot, each time changing when running the function.
}
\description{
Creating an overview map of a random scene of the data to check the results
A follow up after using function 'min_points' and 'outer_points'. Their
output files will be used here.
An overview op of the entire tracks and points would be too messy
so each time running this function, a random spot on your tracks will be mapped to see how it worked out.
The idea is to adjust the parameters used in the functions that created the points.
}
