% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_point.R
\name{outer_point}
\alias{outer_point}
\title{Generates the outer limit of a side of the path as points on the crossprofiles.}
\usage{
outer_point(
  dsm,
  tracks,
  side1 = TRUE,
  export = TRUE,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{GeoPackage file of lines}

\item{side1}{If 'TRUE' side 1 will be returned. If 'FALSE' side 2 will be returned.}

\item{export}{If 'TRUE' both sides will be exported as GeoPackage.}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}
}
\value{
A GeoPackage of points on the left and right side of the path.
}
\description{
Returns the outer limit of the path of ONE side. You can choose the side you want.
Can export two GeoPackages. Each is the outer limit of the path of each side of the path.
}
\details{
TIP: Run this function twice to get both sides into the environment as 'side1_extent' and 'side2_extent' for a better workflow.
Be sure to activate the according parameter.
}
