% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer_points.R
\name{outer_points}
\alias{outer_points}
\title{Generates the outer limit of the path as points on the crossprofiles.}
\usage{
outer_points(
  dsm,
  tracks,
  dist_cross = 1,
  profile_length = 1,
  dist_cross_points = 0.05
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{GeoPackage file of lines}

\item{dist_cross}{Distance between each crossprofile in meter. Defaults to '1'.}

\item{profile_length}{Length of the crossprofile in meter. Defaults to '1'.}

\item{dist_cross_points}{Distance of the points on the crossprofile in meter. Defaults to '0.05'.}
}
\value{
A GeoPackage of points on the left and right side of the path.
}
\description{
Generates the outer limit of the path as points on the crossprofiles.
}
