% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dsm.R
\name{read_dsm}
\alias{read_dsm}
\title{Imports the raster (a Digital Surface Model - dsm).}
\usage{
read_dsm(yourDSMpath)
}
\arguments{
\item{yourDSMpath}{Your path to your GEOTIFF file (.tif)}
}
\value{
the imported dsm file
}
\description{
For better workflow use function as "dsm <- read_dsm(yourDSMpath)"
}
