% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tracks.R
\name{read_tracks}
\alias{read_tracks}
\title{Imports your track data}
\usage{
read_tracks(yourTracksPath)
}
\arguments{
\item{yourTracksPath}{Your path to your GeoPackage file}
}
\description{
This functions helps to prepare your track-data for the other steps.
This step is highly recommended before using the other functions.
For a better workflow use function as "tracks <- read_track(yourTracksPath)"
}
\examples{
tracks <- read_tracks(system.file("geopackage/tracks_d.gpkg", package = "checkRtrack"))
tracks


}
