% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMap.R
\name{checkMap}
\alias{checkMap}
\title{Creates an overview map of a random scene of the data to check the results.}
\usage{
checkMap(
  dsm,
  tracks,
  points,
  morepoints = NULL,
  evenmorepoints = NULL,
  export = FALSE
)
}
\arguments{
\item{dsm}{Digital Surface Model raster file as '.tif'.}

\item{tracks}{Digital Surface Model raster file as '.tif'.}

\item{points}{A Geopackage of Points that you want to plot}

\item{morepoints}{A Geopackage of Points that you want to plot additionally}

\item{evenmorepoints}{A Geopackage of Points that you want to plot additionally}

\item{export}{If 'TRUE', the plot as png will be exported to your wd.}
}
\description{
An overview op of the entire tracks and points would be too messy,
so each time running this function, a random spot on your tracks will be mapped to see how it worked out (so zooming in).
The idea is to adjust the parameters used in the functions that created the points.
}
