% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRtrack-package.R
\docType{package}
\name{checkRtrack-package}
\alias{checkRtrack}
\alias{checkRtrack-package}
\title{checkRtrack: R package to define track widths.}
\description{
You can generate track widths and adjust the process interactively by checking and mapping it. You need digitized tracks and a Digital Surface Model. It is created for the course "Introduction to Programming with R".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geoniesun/checkRtrack/tree/main}
}

}
\author{
\strong{Maintainer}: Leonie Sonntag \email{leonie.sonntag@stud-mail.uni-wuerzburg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
